Add `redirect_from` and split examples in two
